+++ docs/conf.py
@@ -20,7 +20,7 @@
 project = 'imagedl'
 copyright = '2022-2030, Zhenchao Jin'
 author = 'Zhenchao Jin'
-release = '0.3.1'
+release = '0.3.2'
 
 # -- General configuration ---------------------------------------------------