+++ scripts/test_image_clients.py
@@ -8,7 +8,7 @@ WeChat Official Account (微信公众号):
 '''
 from imagedl.modules.sources import (
     BingImageClient, I360ImageClient, YahooImageClient, BaiduImageClient, SogouImageClient, GoogleImageClient, YandexImageClient, PixabayImageClient, 
-    DuckduckgoImageClient, UnsplashImageClient, GelbooruImageClient, SafebooruImageClient, DanbooruImageClient, PexelsImageClient
+    DuckduckgoImageClient, UnsplashImageClient, GelbooruImageClient, SafebooruImageClient, DanbooruImageClient, PexelsImageClient, DimTownImageClient
 )
 
 # bing tests
@@ -66,4 +66,8 @@ client.download(image_infos, num_threadings=1)
 # pexels tests
 client = PexelsImageClient()
 image_infos = client.search('animals', search_limits=10, num_threadings=1)
+client.download(image_infos, num_threadings=1)
+# dimtown tests 
+client = DimTownImageClient()
+image_infos = client.search('美女', search_limits=10, num_threadings=1)
 client.download(image_infos, num_threadings=1)